Compile on Cab
Compiles with the intel compilers.  Does not encounter the malloc issue.

The intel compilers don't understand std::to_string(int) because there are too many equivalent
choices, so a static cast was needed.

+++ include/Path.cc
@@ -6,7 +6,7 @@ std::string multi2d_int_to_string(const multi2d<int> &P){
     if( P.nrows() == 0) return "ZERO";
     for(int r=0; r<P.nrows(); r++){
         for_direction(d){
-            if( P[r][d] != 0) result+=std::to_string(d)+": "+std::to_string(P[r][d])+"; ";
+            if( P[r][d] != 0) result+=std::to_string(static_cast<long long>(d))+": "+std::to_string(static_cast<long long>(P[r][d]))+"; ";
         }
     }
     return result;
